### 2018-07-24 d3dd175
remove font in header because it beaks joss compilation
