### 2018-07-24 #838 @vincentarelbundock
Thanks a lot for working on this. I really appreciate the time an efforts you invest in JOSS!

I added a `LICENSE.txt` file to the repo.

Also, I suspect that the PDF compilation failed because I had added this to the markdown header `CJKmainfont: STSong` (chinese characters didn't render locally on my computer without that line). I have now removed that line. Hopefully this fixes the failed to compile
