### 2018-07-24 #838 @NilsEnevoldsen
@vincentarelbundock Looks like you need to add back `CJKmainfont` if it doesn't break anything.
