### 2018-07-24 daaaecd
add chinese characters back in joss
