### 2018-07-24 #838 @vincentarelbundock
I just tried, but the chinese characters don't show up, even with `CJKmainfont`. If there's no way to render these characters, we can find another usage example to replace that one.
