### 2018-07-25 dc61c71
switching to Gleditsch's other site and fixing unmatched warnings.
