+++ dictionary/get_gw.R
@@ -1,6 +1,6 @@
 get_gw = function(){
 
-  url = "http://privatewww.essex.ac.uk/~ksg/data/iisystem.dat"
+  url = "http://ksgleditsch.com/data/iisystem.dat"
 
   gw = readr::read_tsv(url,
                        col_names = c("gwn", "gwc", "country", "birth", "death"),
@@ -13,20 +13,19 @@ get_gw = function(){
                        ))
 
   gw = gw %>%
-    dplyr::filter(death > as.Date("1945-09-02")) %>%
+    dplyr::filter(
+      !gwc %in% c("TBT", "DRV"),
+      death > as.Date("1945-09-02")
+    ) %>%
     dplyr::group_by(gwn) %>%
     dplyr::slice(which.max(death)) %>%
     dplyr::ungroup() %>%
     dplyr::mutate(
       country = replace(country, gwc == "CDI", "Cote d'Ivoire"),
       country = replace(country, gwc == "PRK", "North Korea"),
-      country.name.en.regex = CountryToRegex(country),
-      country.name.en.regex = replace(country.name.en.regex, gwc == "TBT",
-                                      "tibet"),
-      country.name.en.regex = replace(country.name.en.regex, gwc == "DRV",
-                                      "democratic.republic.of.vietnam"),
-      country.name.en.regex = replace(country.name.en.regex, gwc == "RVN",
-                                      "republic.of.vietnam")) %>%
+      country = replace(country, gwc == "RVN", "Vietnam"),
+      country.name.en.regex = CountryToRegex(country)
+    ) %>%
     dplyr::select(gwn, gwc, country.name.en.regex)
 
   return(gw)
