### 2018-07-26 #848 @whedon
[REVIEW]: countrycode: An R package to convert country names and country codes
**Submitting author:** @vincentarelbundock (<a href="http://orcid.org/0000-0003-2042-7063">Vincent Arel-Bundock</a>)
**Repository:** <a href="https://github.com/vincentarelbundock/countrycode" target ="_blank">https://github.com/vincentarelbundock/countrycode</a>
**Version:** 1.00.0
