### 2018-07-26 #848 @leeper
@andybega The review checklist is here - work through and see if anything is missing or could be improved, flag any relevant issues, and ping me when you've completed your review. It sounds like you're already familiar with the package, so I imagine this should be pretty straightforward. Thanks, again!
