### 2018-08-08 vincentarelbundock/countrycode#188 @andybega
Release tag for JOSS version
Hello, as fortune would have it I'm reviewing your JOSS submission at https://github.com/openjournals/joss-reviews/issues/848. One very minor issue, would you mind tagging the JOSS version as a release? I see that there is a commit (506291a9d5948f70aae28684134ee5f9413c4eba) for it but this would just make it slightly easier to find in case anyone might want to.
