### 2018-08-08 #848 @andybega
@leeper I've gone through the checklist and everything looks good. This package has been around for several years and the core functionality works as advertised. There are extensive tests, the documentation and README are good, and there are clear guidelines for potential contributors.

Not sure if the Chinese character issue in the paper is something that should be resolved in whedon first, but otherwise everything else is in order.
