### 2018-08-08 #848 @leeper
Thanks, @andybega! Much appreciate the quick and thorough review. I agree package is very robust.
