### 2018-08-08 vincentarelbundock/countrycode#188 @vincentarelbundock
Thanks a lot for your work at JOSS. This is a great initiative and I really appreciate those who put in the work!

And thanks for the useful suggestion: I a tag: https://github.com/vincentarelbundock/countrycode/releases/tag/joss
