### 2018-08-08 #848 @vincentarelbundock
Thanks @andybega for your nice words and speedy review. I really appreciate it!

@leeper  here's the Zenodo DOI:  https://doi.org/10.5281/zenodo.1341914

With respect to https://github.com/openjournals/whedon/pull/30, I'm happy to do whatever's easiest for you. Given the nature of the package, I thought a non-latin example was neat, but it would be trivial to change the output to Italian or some such.

Just let me know what you prefer.

Thanks again for your work on JOSS!
