### 2018-08-08 #848 @leeper
@vincentarelbundock Let's let @arfon weigh in on how to proceed. Everything else is ready to go, so that may be a trivial fix (Arfon: https://github.com/openjournals/whedon/pull/30). If it turns out it's not, then updating the paper might be the easier way to go.
