### 2018-08-08 vincentarelbundock/countrycode#188 @vincentarelbundock
https://github.com/vincentarelbundock/countrycode/releases/tag/1.00.0-joss

cleaner version
