### 2018-08-09 #848 @arfon
> If it turns out it's not, then updating the paper might be the easier way to go.

I tried taking https://github.com/openjournals/whedon/pull/30 for a spin and it's caused a whole bunch of problems with other packages (weird clashing config errors in packages we depend upon) which look like they are going to be non-trivial to fix.

Basically, if there is a workaround without https://github.com/openjournals/whedon/pull/30 then I think we should pursue it.
