### 2018-08-09 bf00bc0
remove chinese characters from joss paper (their latex template does not support)
