### 2018-08-09 #848 @vincentarelbundock
@arfon Thanks a lot for looking into this. No worries. I switched from Chinese to Finnish, and the proof looks good to me.

Pinging @NilsEnevoldsen and @cjyetman in case they want to take a last look.
