### 2018-08-09 #848 @leeper
Looks good to me. @arfon I think it's back to you.
