### 2018-08-09 #848 @vincentarelbundock
Thanks @arfon , this is great news!

I really appreciate all of your work on JOSS. This is a fantastic initiative. I'm now in the reviewer pool, so feel free to hit me up.
